+++ extensions/things/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Quick To-do Improvements] - {PR_MERGE_DATE}
 
 - Preference to disable date parsing for todo classification, sending all todo(s) to inbox.
+- The title input will still be parsed for assigning deadlines and list name followed with '#' (case-insensitive).
 
 ## [Fix Generate Checklist with AI] - 2024-07-30